Add Clion project

+++ src/main.cpp
@@ -1,7 +1,6 @@
 #include <Arduino.h>
-#include <SoftwareSerial.h>
 
-int relay1 = 2;
+uint8_t relay1 = 2;
 
 byte serialA;
 void setup() {
@@ -12,7 +11,7 @@ void setup() {
 
 void loop() {
     if (Serial.available() > 0) {
-	serialA = Serial.read();Serial.println(serialA);
+	serialA = (byte) Serial.read();Serial.println(serialA);
     }
     switch (serialA) {
     case 1:
